+++ rapid_react.py
@@ -1,29 +1,29 @@
 from dataclasses import dataclass
 from io import TextIOWrapper
 import json
-from models import Element
+from models import Element, GameElementState
 
 
 
 @dataclass
-class GameElementState:
-    '''Represents the current state of the game'''
+class RapidReactGameElementState(GameElementState):
+    '''Represents the current state of the rapid react game'''
     red_cargo: list[Element]
     blue_cargo: list[Element]
-    misc : list[Element]
+    misc: list[Element]
 
     @staticmethod
-    def read(file: TextIOWrapper) -> 'GameElementState':
+    def read(file: TextIOWrapper) -> 'RapidReactGameElementState':
         '''Returns the current state of the game'''
         raw = file.read()
         raw = raw.strip()
         raw = json.loads(raw)
-        elements = []
+        elements: list[Element] = []
         for raw_object in raw['objects']:
             elements.append(Element.from_json(raw_object))
-        red_cargo = []
-        blue_cargo = []
-        misc = []
+        red_cargo: list[Element] = []
+        blue_cargo: list[Element] = []
+        misc: list[Element] = []
         for element in elements:
             if element.element_type is None:
                 misc.append(element)
@@ -35,7 +35,7 @@ class GameElementState:
                 misc.append(element)
         red_cargo.sort(key=lambda i: i.identifier)
         blue_cargo.sort(key=lambda i: i.identifier)
-        return GameElementState(red_cargo, blue_cargo, misc)
+        return RapidReactGameElementState(red_cargo, blue_cargo, misc)
 
     def __str__(self) -> str:
         return f"{[str(item) for item in self.red_cargo]}\n" \
